requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.5
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-extensions==3.1.3
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-widget-tweaks==1.4.8
filelock==3.0.12
mysql==0.0.3
mysqlclient==2.0.3
platformdirs==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.6.0
virtualenvwrapper-win==1.2.6
